Replacing pyparsing parser with lex/yacc parser
