+++ source/pyconform/parsing.py
@@ -1,186 +1,276 @@
 """
-Parsing Module
+Parsing Module - NEW Based on PLY
 
 This module defines the necessary elements to parse a string variable definition
 into the recognized elements that are used to construct an Operation Graph.
 
-Copyright 2017, University Corporation for Atmospheric Research
+Copyright 2017-2018, University Corporation for Atmospheric Research
 LICENSE: See the LICENSE.rst file for details
 """
 
-from numpy import index_exp
-from pyparsing import nums, alphas, alphanums, oneOf, delimitedList, opAssoc, operatorPrecedence
-from pyparsing import Word, Combine, Forward, Suppress, Group, Optional, ParserElement
-from pyparsing import Literal, CaselessLiteral, QuotedString
+from ply import lex, yacc
 
-# To improve performance
-ParserElement.enablePackrat()
+tokens = ('UINT', 'UFLOAT', 'STRING', 'NAME', 'POW', 'EQ', 'LEQ', 'GEQ')
+literals = ('*', '/', '+', '-', '<', '>', '=', ',', ':', '(', ')', '[', ']')
+t_ignore = ' \t'
 
-#===================================================================================================
-# ParsedFunction
-#===================================================================================================
-class ParsedFunction(object):
-    """
-    A parsed function string-type
-    """
-
-    def __init__(self, tokens):
-        token = tokens[0]
-        self.key = token[0]
-        self.args = []
-        self.kwds = {}
-        for t in token[1:]:
-            if isinstance(t, tuple):
-                self.kwds[t[0]] = t[1]
-            else:
-                self.args.append(t)
-        self.args = tuple(self.args)
-    def __repr__(self):
-        clsname = self.__class__.__name__
-        argstr = ','.join('{!r}'.format(a) for a in self.args)
-        kwdstr = ','.join('{}={!r}'.format(k, self.kwds[k]) for k in self.kwds)
-        if len(self.args) > 0:
-            paramstr = '{},{}'.format(argstr, kwdstr) if len(self.kwds) > 0 else argstr
-        else:
-            paramstr = kwdstr if len(self.kwds) > 0 else ''
-        return ("<{} {}({}) ('{}') at {}>").format(clsname, self.key, paramstr, str(self), hex(id(self)))
-    def __str__(self):
-        argstr = ','.join('{!s}'.format(a) for a in self.args)
-        kwdstr = ','.join('{}={!s}'.format(k, self.kwds[k]) for k in self.kwds)
-        if len(self.args) > 0:
-            paramstr = '{},{}'.format(argstr, kwdstr) if len(self.kwds) > 0 else argstr
-        else:
-            paramstr = kwdstr if len(self.kwds) > 0 else ''
-        return "{}({!s})".format(self.key, paramstr)
-    def __eq__(self, other):
-        return ((type(self) == type(other)) and (self.key == other.key) and
-                (self.args == other.args) and (self.kwds == other.kwds))
-        
-        
-#===================================================================================================
-# ParsedUniOp
-#===================================================================================================
-class ParsedUniOp(ParsedFunction):
-    """
-    A parsed unary-operator string-type
-    """
-    def __str__(self):
-        return "({}{!s})".format(self.key, self.args[0])
+t_NAME = r'[a-zA-Z_][a-zA-Z0-9_]*'
+t_POW = r'\*\*'
+t_LEQ = r'<='
+t_GEQ = r'>='
+t_EQ = r'=='
 
 
-#===================================================================================================
-# ParsedBinOp
-#===================================================================================================
-class ParsedBinOp(ParsedFunction):
-    """
-    A parsed binary-operator string-type
-    """
-    def __str__(self):
-        return "({!s}{}{!s})".format(self.args[0], self.key, self.args[1])
+def t_UFLOAT(t):
+    r'(([0-9]+\.[0-9]*|[0-9]*\.[0-9]+)([eE][+-]?[0-9]+)?|[0-9]+[eE][+-]?[0-9]+)'
+    t.value = float(t.value)
+    return t
 
 
-#===================================================================================================
-# ParsedVariable
-#===================================================================================================
-class ParsedVariable(ParsedFunction):
-    """
-    A parsed variable string-type
-    """
-    def __init__(self, tokens):
-        super(ParsedVariable, self).__init__(tokens)
-        self.args = index_exp[self.args] if len(self.args) > 0 else ()
-    def __repr__(self):
-        return "<{} '{}' at {}>".format(self.__class__.__name__, str(self), hex(id(self)))
-    def __str__(self):
-        if len(self.args) == 0:
-            strargs = ''
-        else:
-            strargs = str(list(self.args))
-        return "{}{}".format(self.key, strargs)
+def t_UINT(t):
+    r'[0-9]+'
+    t.value = int(t.value)
+    return t
 
 
-#===================================================================================================
-# Operator Parser Functions
-#===================================================================================================
+def t_STRING(t):
+    r'"([^"\\]*(\\.[^"\\]*)*)"|\'([^\'\\]*(\\.[^\'\\]*)*)\''
+    t.value = t.value[1:-1]
+    return t
 
-# Negation operator
-def _negop_(tokens):
-    op, val = tokens[0]
-    if op == '+':
-        return val
+
+def t_error(t):
+    raise TypeError('Unexpected string: {!r}'.format(t.value))
+
+
+lex.lex()
+
+
+from collections import namedtuple
+
+
+def ind_str(index):
+    if isinstance(index, slice):
+        ind_list = [index.start, index.stop, index.step]
+        _str = ':'.join('' if i is None else str(i) for i in ind_list)
+        return ':' if _str == '::' else _str
     else:
-        return ParsedUniOp([[op, val]])
+        return str(index)
 
-# Binary Operators
-def _binop_(tokens):
-    left, op, right = tokens[0]
-    return ParsedBinOp([[op, left, right]])
 
-#===================================================================================================
-# MAIN PARSER (DEFINED AT MODULE LEVEL TO MAKE A SINGLETON)
-#===================================================================================================
+UniOpType = namedtuple('UniOpType', ['key', 'args'])
+UniOpType.__new__.__defaults__ = (None, [])
+UniOpType.__str__ = lambda self: '({}{})'.format(self.key, self.args[0])
 
-# INTEGERS: Just any word consisting only of numbers
-_INT_ = Word(nums)
-_INT_.setParseAction(lambda t: int(t[0]))
+BinOpType = namedtuple('BinOpType', ['key', 'args'])
+BinOpType.__new__.__defaults__ = (None, [])
+BinOpType.__str__ = lambda self: '({}{}{})'.format(
+    self.args[0], self.key, self.args[1])
 
-# FLOATS: More complicated... can be decimal format or exponential
-#         format or a combination of the two
-_DEC_FLT_ = (Combine(Word(nums) + '.' + Word(nums)) |
-             Combine(Word(nums) + '.') |
-             Combine('.' + Word(nums)))
-_EXP_FLT_ = (Combine(CaselessLiteral('e') + Optional(oneOf('+ -')) + Word(nums)))
-_FLOAT_ = (Combine(Word(nums) + _EXP_FLT_) | Combine(_DEC_FLT_ + Optional(_EXP_FLT_)))
-_FLOAT_.setParseAction(lambda t: float(t[0]))
+FuncType = namedtuple('FuncType', ['key', 'args', 'kwds'])
+FuncType.__new__.__defaults__ = (None, [], {})
+FuncType.__str__ = lambda self: '{}({})'.format(
+    self.key, ','.join([str(a) for a in self.args] +
+                       ['{}={}'.format(k, self.kwds[k]) for k in self.kwds]))
 
-# QUOTED STRINGS: Any words between quotations
-_QSTR_ = QuotedString('"', escChar='\\')
+VarType = namedtuple('VarType', ['key', 'ind'])
+VarType.__new__.__defaults__ = (None, [])
+VarType.__str__ = lambda self: '{}{}'.format(
+    self.key, '' if len(self.ind) == 0 else '[{}]'.format(','.join([ind_str(a) for a in self.ind])))
 
-# String _NAME_s ...identifiers for function or variable _NAME_s
-_NAME_ = Word(alphas + "_", alphanums + "_")
 
-# FUNCTIONS: Function arguments can be empty or any combination of
-#            ints, _FLOAT_, variables, and even other functions.  Hence,
-#            we need a Forward place-holder to start...
-_EXPR_PARSER_ = Forward()
+precedence = (('left', 'EQ'),
+              ('left', '<', '>', 'LEQ', 'GEQ'),
+              ('left', '+', '-'),
+              ('left', '*', '/'),
+              ('right', 'NEG', 'POS'),
+              ('left', 'POW'))
 
-# Named Arguments
-_KWDS_ = Group( _NAME_ + Suppress('=') + (_QSTR_ | _EXPR_PARSER_) )
-_KWDS_.setParseAction(lambda t: tuple(*t))
 
-# Functions
-_FUNC_ = Group(_NAME_ + (Suppress('(') + Optional(delimitedList(_QSTR_ | _KWDS_ | _EXPR_PARSER_)) +
-                         Suppress(')')))
-_FUNC_.setParseAction(ParsedFunction)
+def p_array_like(p):
+    """
+    array_like : UFLOAT
+    array_like : UINT
+    array_like : function
+    array_like : variable
+    """
+    p[0] = p[1]
 
-# VARIABLE NAMES: Can be just string _NAME_s or _NAME_s with blocks
-#                 of indices (e.g., [1,2,-4])
-_INDEX_ = Combine(Optional('-') + Word(nums))
-_INDEX_.setParseAction(lambda t: int(t[0]))
 
-_IDX_OR_NONE_ = Optional(_INDEX_)
-_IDX_OR_NONE_.setParseAction(lambda t: t[0] if len(t) > 0 else [None])
+def p_array_like_group(p):
+    """
+    array_like : '(' array_like ')'
+    """
+    p[0] = p[2]
 
-# _ISLICE_ = _IDX_OR_NONE_ + Optional(Suppress(':') + _IDX_OR_NONE_ + Optional(Suppress(':') + _IDX_OR_NONE_))
-# _ISLICE_.setParseAction(lambda t: slice(*t) if len(t) > 1 else t[0])
-_ISLICE_ = delimitedList(_IDX_OR_NONE_, delim=':')
-_ISLICE_.setParseAction(lambda t: slice(*t) if len(t) > 1 else t[0])
 
-_VARIABLE_ = Group(_NAME_ + Optional(Suppress('[') + delimitedList(_ISLICE_ | _INDEX_) + Suppress(']')))
-_VARIABLE_.setParseAction(ParsedVariable)
+def p_function_with_arguments_and_keywords(p):
+    """
+    function : NAME '(' argument_list ',' keyword_dict ')'
+    """
+    p[0] = FuncType(p[1], p[3], p[5])
 
-# Expression parser
-_EXPR_PARSER_ << operatorPrecedence(_FLOAT_ | _INT_ | _FUNC_ | _VARIABLE_,
-                                    [(Literal('**'), 2, opAssoc.RIGHT, _binop_),
-                                     (oneOf('+ -'), 1, opAssoc.RIGHT, _negop_),
-                                    (Literal('/'), 2, opAssoc.RIGHT, _binop_),
-                                    (Literal('*'), 2, opAssoc.RIGHT, _binop_),
-                                    (Literal('-'), 2, opAssoc.RIGHT, _binop_),
-                                    (Literal('+'), 2, opAssoc.RIGHT, _binop_)])
 
-#===================================================================================================
+def p_function_with_arguments_only(p):
+    """
+    function : NAME '(' argument_list ')'
+    """
+    p[0] = FuncType(p[1], p[3], {})
+
+
+def p_function_with_keywords_only(p):
+    """
+    function : NAME '(' keyword_dict ')'
+    """
+    p[0] = FuncType(p[1], [], p[3])
+
+
+def p_argument_list_append(p):
+    """
+    argument_list : argument_list ',' argument
+    """
+    p[0] = p[1] + [p[3]]
+
+
+def p_single_item_argument_list(p):
+    """
+    argument_list : argument
+    argument_list : 
+    """
+    p[0] = [p[1]] if len(p) > 1 else []
+
+
+def p_argument(p):
+    """
+    argument : array_like
+    argument : STRING
+    """
+    p[0] = p[1]
+
+
+def p_keyword_dict_setitem(p):
+    """
+    keyword_dict : keyword_dict ',' NAME '=' argument
+    """
+    p[1][p[3]] = p[5]
+    p[0] = p[1]
+
+
+def p_single_item_keyword_dict(p):
+    """
+    keyword_dict : NAME '=' argument
+    """
+    p[0] = {p[1]: p[3]}
+
+
+def p_variable(p):
+    """
+    variable : NAME '[' index_list ']'
+    variable : NAME
+    """
+    indices = p[3] if len(p) > 3 else []
+    p[0] = VarType(p[1], indices)
+
+
+def p_index_list_append(p):
+    """
+    index_list : index_list ',' index
+    """
+    p[0] = p[1] + [p[3]]
+
+
+def p_single_item_index_list(p):
+    """
+    index_list : index
+    """
+    p[0] = [p[1]]
+
+
+def p_index(p):
+    """
+    index : slice
+    index : array_like
+    """
+    p[0] = p[1]
+
+
+def p_slice(p):
+    """
+    slice : slice_argument ':' slice_argument ':' slice_argument
+    slice : slice_argument ':' slice_argument
+    """
+    p[0] = slice(*p[1::2])
+
+
+def p_slice_argument(p):
+    """
+    slice_argument : array_like
+    slice_argument : 
+    """
+    p[0] = p[1] if len(p) > 1 else None
+
+
+def p_expression_unary(p):
+    """
+    array_like : '-' array_like %prec NEG
+    array_like : '+' array_like %prec POS
+    """
+    if p[1] == '+':
+        p[0] = p[2]
+    elif p[1] == '-':
+        if isinstance(p[2], (UniOpType, BinOpType, FuncType, VarType)):
+            p[0] = UniOpType(p[1], [p[2]])
+        else:
+            p[0] = -p[2]
+
+
+def p_expression_binary(p):
+    """
+    array_like : array_like POW array_like
+    array_like : array_like '-' array_like
+    array_like : array_like '+' array_like
+    array_like : array_like '*' array_like
+    array_like : array_like '/' array_like
+    array_like : array_like '<' array_like
+    array_like : array_like '>' array_like
+    array_like : array_like LEQ array_like
+    array_like : array_like GEQ array_like
+    array_like : array_like EQ array_like
+    """
+    if (isinstance(p[1], (UniOpType, BinOpType, FuncType, VarType)) or
+            isinstance(p[3], (UniOpType, BinOpType, FuncType, VarType))):
+        p[0] = BinOpType(p[2], [p[1], p[3]])
+    elif p[2] == '**':
+        p[0] = p[1] ** p[3]
+    elif p[2] == '-':
+        p[0] = p[1] - p[3]
+    elif p[2] == '+':
+        p[0] = p[1] + p[3]
+    elif p[2] == '*':
+        p[0] = p[1] * p[3]
+    elif p[2] == '/':
+        p[0] = p[1] / p[3]
+    elif p[2] == '<':
+        p[0] = p[1] < p[3]
+    elif p[2] == '>':
+        p[0] = p[1] > p[3]
+    elif p[2] == '<=':
+        p[0] = p[1] <= p[3]
+    elif p[2] == '>=':
+        p[0] = p[1] >= p[3]
+    elif p[2] == '==':
+        p[0] = p[1] == p[3]
+
+
+def p_error(p):
+    raise TypeError('Parsing error at {!r}'.format(p.value))
+
+
+yacc.yacc()
+
+
+#=========================================================================
 # Function to parse a string definition
-#===================================================================================================
-
+#=========================================================================
 def parse_definition(strexpr):
-    return _EXPR_PARSER_.parseString(strexpr)[0]
+    return yacc.parse(strexpr)  # @UndefinedVariable
